fix remove room

+++ src/Service/RemoveRoomService.php
@@ -63,10 +63,12 @@ class RemoveRoomService
             }
             $this->em->persist($room);
             $this->em->flush();
-            $callerRoom = $room->getCallerRoom();
-            $this->em->remove($callerRoom);
-            $this->em->flush();
 
+            if ($room->getCallerRoom()){
+                $callerRoom = $room->getCallerRoom();
+                $this->em->remove($callerRoom);
+                $this->em->flush();
+            }
 
         } catch (\Exception $e) {
             $this->logger->error($e->getMessage());
